imagehelper.js converted to .ts
