+++ vue-project/.eslintrc.cjs
@@ -1,14 +1,14 @@
-/* eslint-env node */
-require('@rushstack/eslint-patch/modern-module-resolution')
+// /* eslint-env node */
+// require('@rushstack/eslint-patch/modern-module-resolution')
 
-module.exports = {
-  root: true,
-  'extends': [
-    'plugin:vue/vue3-essential',
-    'eslint:recommended',
-    '@vue/eslint-config-prettier/skip-formatting'
-  ],
-  parserOptions: {
-    ecmaVersion: 'latest'
-  }
-}
+// module.exports = {
+//   root: true,
+//   'extends': [
+//     'plugin:vue/vue3-essential',
+//     'eslint:recommended',
+//     '@vue/eslint-config-prettier/skip-formatting'
+//   ],
+//   parserOptions: {
+//     ecmaVersion: 'latest'
+//   }
+// }
